refactor: move search helpers to app router
